Outputs both table details and incident details
Significant update to gvascrape. In addition to saving the table details to a file, it now saves incident details to a separate file (e.g. the content of http://www.gunviolencearchive.org/incident/765559). The two files can be joined on incident_id.
I extended the command line interface to accept two file names and the nextURL parameter. NextUrl allows you to scrape different year sets on the gunviolencearchive.